Moved ts.Node logging utilities to utils.ts
